+++ kifi-backend/modules/shoebox/app/com/keepit/common/mail/MailModule.scala
@@ -3,7 +3,7 @@ package com.keepit.common.mail
 import com.google.inject.{Provides, Singleton}
 import play.api.Play._
 import com.keepit.inject.AppScoped
-import com.keepit.common.healthcheck.{AirbrakeNotifier, HealthcheckMailSender, LocalHealthcheckMailSender}
+import com.keepit.common.healthcheck.{AirbrakeNotifier, SystemAdminMailSender, LocalSystemAdminMailSender}
 import net.codingwell.scalaguice.ScalaModule
 import com.keepit.common.plugin.SchedulingProperties
 import com.keepit.common.actor.ActorInstance
@@ -17,7 +17,7 @@ case class ProdMailModule() extends MailModule {
     bind[LocalPostOffice].to[ShoeboxPostOfficeImpl]
     bind[MailSenderPlugin].to[MailSenderPluginImpl].in[AppScoped]
     bind[MailToKeepPlugin].to[MailToKeepPluginImpl].in[AppScoped]
-    bind[HealthcheckMailSender].to[LocalHealthcheckMailSender]
+    bind[SystemAdminMailSender].to[LocalSystemAdminMailSender]
   }
 
   @Singleton
@@ -36,7 +36,7 @@ case class DevMailModule() extends MailModule {
   def configure() {
     bind[LocalPostOffice].to[ShoeboxPostOfficeImpl]
     bind[MailSenderPlugin].to[MailSenderPluginImpl].in[AppScoped]
-    bind[HealthcheckMailSender].to[LocalHealthcheckMailSender]
+    bind[SystemAdminMailSender].to[LocalSystemAdminMailSender]
   }
 
   @Provides
